Moved set measurement CLI to dedicated module
